Move createModal and haveRival into own files
